requirements.txt
aiogram
yfinance
python-dotenv